+++ app/controllers/phone_callbacks_controller.rb
@@ -1,40 +1,33 @@
 class PhoneCallbacksController < ApplicationController
   before_action :require_user
-  before_action :check_if_super_user?, only: %i[approvals]
-  before_action :check_if_sys_admin?, only: %i[verifications]
+  before_action :check_if_super_user?, only: [:approvals]
+  before_action :check_if_sys_admin?, only: [:verifications]
   before_action :set_per_page
+  before_action :set_filter_option
 
   def approvals
-    if params[:search].present?
-      @certificate_orders = CertificateOrder.includes(:messages, :registrants, :certificate_contents)
-        .where(ref: params[:search])
-        .joins{ sub_order_items.product_variant_item.product_variant_group.variantable(Certificate) }
-        .where(registrants: { phone_number_approved: false })
-        .where(messages: { subject: 'Request for approving Phone Number' } )
+    if params[:search].present? || params[:filter].present?
+      @certificate_orders = find_pending_approvals
       if @certificate_orders.empty?
         render :approvals
       else
-        @certificate_orders = @certificate_orders.paginate(page: params[:page], per_page: @per_page || 10)
+        @certificate_orders = @certificate_orders.paginate(page: params[:page], per_page: @per_page)
       end
     else
-      @certificate_orders = find_pending_approvals.paginate(page: params[:page], per_page: @per_page || 10)
+      @certificate_orders = find_pending_approvals.paginate(page: params[:page], per_page: @per_page)
     end
   end
 
   def verifications
-    if params[:search].present?
-      @certificate_orders = CertificateOrder.includes(:certificate_order_tokens, :registrants, :certificate_contents)
-      .where(ref: params[:search])
-      .where.not(contacts: { contactable_id: nil })
-      .where.not(certificate_contents: { workflow_state: 'issued' })
-      .where(certificate_order_tokens: { status: 'pending', callback_type: 'manual', callback_method: 'call' })
+    if params[:search].present? || params[:filter].present?
+      @certificate_orders = find_pending_verifications
       if @certificate_orders.empty?
         render :verifications
       else
-        @certificate_orders = @certificate_orders.paginate(page: params[:page], per_page: @per_page || 10)
+        @certificate_orders = @certificate_orders.paginate(page: params[:page], per_page: @per_page)
       end
     else
-      @certificate_orders = find_pending_verifications.paginate(page: params[:page], per_page: @per_page || 10)
+      @certificate_orders = find_pending_verifications.paginate(page: params[:page], per_page: @per_page)
     end
   end
 
@@ -54,7 +47,18 @@ class PhoneCallbacksController < ApplicationController
   private
 
   def set_per_page
-    @per_page = params[:number_rows]
+    preferred_row_count = current_user.try('preferred_cert_order_row_count')
+    @per_page = params[:number_rows] || preferred_row_count
+    CertificateOrder.per_page = @per_page if CertificateOrder.per_page != @per_page
+    current_user&.update_attribute('preferred_cert_order_row_count', @per_page) if @per_page != preferred_row_count
+  end
+
+  def set_filter_option
+    if (params[:filter].present? && params[:filter] == 'Filter By')
+      @filter = 'Filter By'
+    else
+      @filter = params[:filter]
+    end
   end
 
   def phone_callback_log_params
@@ -70,16 +74,27 @@ class PhoneCallbacksController < ApplicationController
   end
 
   def find_pending_approvals
-    @certificate_orders = CertificateOrder.includes(:messages, :registrants, :certificate_contents)
+    cos = CertificateOrder.includes(:messages, :registrants, :certificate_contents)
       .joins{ sub_order_items.product_variant_item.product_variant_group.variantable(Certificate) }
       .where(registrants: { phone_number_approved: false })
       .where(messages: { subject: 'Request for approving Phone Number' } )
+    search_or_filter(cos)
   end
 
   def find_pending_verifications
-    @certificate_orders = CertificateOrder.includes(:certificate_order_tokens, :registrants, :certificate_contents)
+    cos = CertificateOrder.includes(:certificate_order_tokens, :registrants, :certificate_contents)
       .where.not(contacts: { contactable_id: nil })
       .where.not(certificate_contents: { workflow_state: 'issued' })
       .where(certificate_order_tokens: { status: 'pending', callback_type: 'manual', callback_method: 'call' })
+    search_or_filter(cos)
+  end
+
+  def search_or_filter(cert_orders)
+    if params[:search].present?
+      cert_orders = cert_orders.where(ref: params[:search])
+    elsif params[:filter].present? && params[:filter] != 'Filter By'
+      cert_orders = cert_orders.where("certificates.product = ?", "#{params[:filter]}")
+    end
+    cert_orders
   end
 end