+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ### PENDING RELEASE (TBD)
 
+- Add product filtering option for phone callback approvals and verifications page.
 - Assign CA profile to Business Identity Enterprise product
 - Fix Url Callback function
 - Fix creating 2 SslAccount objects when creating a User account